Add user email to the search index.
ldap and email are technically different in a few ways. Email not only includes the domain (@linkedin.com), but the user part of it may be different than ldap. Generally emails are username@domain; at LinkedIn ldaps are generally usernames truncated to 8 characters.

For the sake of being technically correct; also index emails so that if clients want to search email, they can search email rather than ldap.