Add support for customizing topic names via a convention.
Requested by a few people in OS.

Companies probably need to pick _ or . as a delimiter and we hard coded _

TODO to finish the implmentation for v5. For right now v5 is LI only and unfinished. Getting this in for v4 so it is useful to other companies now.

TODO BEFORE SUBMIT - make configurable via spring
TODO BEFORE SUBMIT - decide if this needs to be an internal RB first
TODO BEFORE SUBMIT - see where else we can use this (jobs, where else?)