Start using spotless to format markdown.
Some of our markdown is quite hard to read in the editor. This helps make it more readable (imo) but also means we don't need to have formatting debates; just run spotlessApply.

This is now enforced with a GitHub workflow. If it fails run spotlessApply and commit those changes.

We don't auto format in a hook because spotless isn't infallible; please check that the formatting looks good before submitting.