ci(docker): disable GitHub Docker registry
The GitHub Docker registry (docker.pkg.github.com) is deprecated and
will be removed. We're already starting to see build flakiness due to
the instability of the image push operations, so this PR disables it
entirely.

Given that almost all DataHub users rely on the Docker Hub images and
almost nobody uses the GitHub Docker registry images, we should be clear
to simply disable them. As such, I'm not migrating to the newer GitHub
Container registry (ghcr.io) right now. Additionally, I suggest that we
also remove the images currently present on GitHub, or the `latest` tags
on those images at a minimum.