feat(ingestion/superset): add base_uri to config
Motivation: Superset can have complicated auth layers preventing direct
access through the main URI.  One way around this is to run a simpler,
private instance and use that for ingestion.  The problem is then that
the URIs in the links are pointing to the private instance.

Change: This change adds a base_uri parameter to the SupersetConfig
source configuration so that connect_uri can be where we ingest from and
base_uri can be used in links for dashboards, graphs, etc.  When absent,
this parameter defaults to connect_uri.