fix(ingestion): add fallback  to trino
Trino `get_table_comment` queries `<table>$properties`. This special
table is only present when using the hive connector.

In order to fix the ingestion for errorring when not using the hive
connector, the original `trino-python-client` behaviour was added as
fallback.

This is a first idea on a quick win to tackle
[ISSUE-5790](https://github.com/datahub-project/datahub/issues/5790)