Update data_platforms.json
SAP Hana (SAP is the Company behind SAP Hana).

Reason 1)
Everywhere (directory, python scripts and so on) is only used "hana"
Cause of consistency it should be "Hana" in the Navigation.

Reason 2)
"SAP Hana" is actually in the Alphabetic order after "G" (S after G, not really :-) cause internally it is sorted for the word "Hana"

BUT
The best would be to use the Name "SAP Hana". Cause this is the used Name in the IT industry. But I dont know for what the "name:" Attribute is used for. Only for sorting or are there other dependencies?