perf(ingest): speed up MCPW.validate()
In my local tests with an MCPW with a dataset schema of 3000 columns,
this reduced runtime from ~676 ms to ~189 ms, which is a 3-4x
improvement. It seems to scale linearly with the size of the MCPW
object, but I didn't test that extensively.

Note that almost 200ms is still pretty bad for basic data validation.
With the changes from https://github.com/acryldata/avro_gen/pull/18, I
was able to get it down to ~130 ms.