feat(ingest): unbundle airflow plugin emitter dependencies
As described at length in
<https://github.com/datahub-project/datahub/blob/v0.10.0/metadata-ingestion/setup.py#L62>
the current state of platform compatibility for the confluent_kafka is
complicated.  It can be a source of installation incompatibility and
general friction.  To make the airflow plugin easier to include in
resolvable dependencies sets across a variety of platforms, unbundle the
datahub-kafka emitter into extras_require.