fix: SchemaDescriptionField 'read-more' doesn't affect table height
`SchemaDescriptionField` is a component used to display descriptions in
`SchemaTable` via `useDescriptionRenderer`.

When `SchemaDescriptionField` expands, `SchemaTable` (antd `Table`)
doesn't know to adjust for the change in height.  This is relevant
when the table starts out shorter than the expanded row, e.g. a table
with only a couple rows, where one of those rows has a long description.

With the expansion state only in SchemaDescriptionField, a change in state
doesn't result in a re-render.

The solution I've come up with is to make `SchemaDescriptionField` a
controlled component wrt expansion, with the state living in
`useDescriptionRenderer`.

`SchemaDescriptionField` is also used by `TableOfMlFeatures`

Related: https://github.com/datahub-project/datahub/issues/7528