fix(metadata ingestion): fix incorrect dot interpretation
Bug: all dataset of all platforms were splitted by dots and the last element of a list was considered as an actual table name.
Fix: add if-statement to check if platform instance is BigQuery, since PostgreSQL and some other DBMSs allow to use dots in table names.
Not the cleanest solution for this behaviour, obviously, so feel free to provide me with an information on how to apply this change in a more consise way.
Consider to move the normalization block to "bigquery_v2/bigquery.py", but I'm not able to test this using actual BQ table.