fix(ingestion/redash): Limit size of RedashSourceReport
Addressing the problem described in #9575

For large Redash deployments, the `filtered` and `timing` fields can grow to be very
large. By using the LossyList and LossyDict data structures in a similar way as used
in other SourceReports, we will limit the number of lines printed during an ingestion
run which makes the logs more usable and improves performance for ingestion against
Redash deployments with a large number of queries filtered.