feat(gms): add ingestProposalBatch endpoint
In my local tests of emitting 1000 MCPs, it was approximately 7x faster
to emit MCPs in batches of 100 instead of emitting one at a time.

This will need to be an opt-in mechanism in the rest sink until the backend
GMS API has been released and stabilized. I have not built that yet.