fix(ingest/dbt): skip CLL parsing if not needed
If column lineage is disabled, we don't need to call _parse_cll to parse
`node.compiled_code` with sqlglot.

If this fails, we get spurious "Failed to generate any CLL lineage"
messages in the logs.

An exception is the case where `added_to_schema_resolver` is False, as
in this case the parsed code is used to infer the schema fields.