feat(ingest/dbt): add reformat_compiled_code option
Calling `try_format_query` on `compiled_code` sometimes does not work,
because sqlglot's generator is recursive, and sufficiently complex
queries can exceed python's recursion limit.

As a workaround, we can just skip the sqlglot formatting and take
the value as it is formatted in the DBT manifest.

Switching off this option is less obtrusive than switching off
`ignore_compiled_code`