fix(ingest): remove full asyncio from check_upgrade
Our previous usage of asyncio was broadly incorrect (it blocked the main
event loop).

It would also not guarantee that ingestion code runs on the main thread.
We want that guarantees, since it enables us to implement timeouts using signals and overall
helps ensure that we get the guarantees we want.

A side benefit is that this might help us simplify the Okta source in
the future.