feat(iceberg source): add config to ignore errors
I've found that having the iceberg source fail on an exception was
reducing the effectiveness of our source ingestion, especially when we don't care.

This change adds a config to the iceberg source to allow you to ignore any
exceptions that arise when DataHub parses iceberg tables. This allows the user
to decide whether they want to be interrupted by errors.