build(coverage): fix carry-forward coverage
Code coverage carry-forward is tied to flags (used to present
modules/subprojects) and code coverage for a flag can be carried forward if
ci optimizations skips a module build. So created a module `flag` for each
group of projects that the ci builds together. This roughly corresponds to
the groups defined in ci-optimization action.