fix(ingest): avoid multiprocessing "fork" start method
We use threads in a variety of places within our CLI. The multiprocessing "fork" start method is not safe to use with threads.
MacOS and Windows already default to "spawn", and Linux will as well starting in Python 3.14.
See https://docs.python.org/3/library/multiprocessing.html#contexts-and-start-methods

This fixes a bug where the classification module would deadlock when
using more than one worker process.

Eventually it may make sense to use "forkserver" as the default where available, but we can revisit that in the future.