Fix platform instance support on Druid ingestion
If the platform instance was specified in a Druid ingestion, it was always
set twice, because it was overrided in the get_identifier, and then added
again down the line.
This may break existing users that already rely on the broken
platform instance implementation, but they could still manually set
it to the old pattern in their dhub file.