combine copy and rename into single task
HAving these as two separate tasks caused copy to always be out-of-date when
rename is run, and rename to be out-of-date when rename is done. Gradle
does not like inplace updates of any files. By combining it, the copy+rename
now appear to produce a new artificat that will be up-to-date if no sources
have changed. This caused lot of downstream jars to get rebuilt since this jar
is always out-of-dat and rebuilt everytime.