revert(ci): revert datahub-wheel build changes
In #12722 and https://github.com/datahub-project/datahub/pull/12730 I
attempted to make the cloudflare pages builds work on PRs from forks. It
turns out cloudflare's pages-action had issues with PRs from forks, and
we started seeing errors (e.g.
https://github.com/datahub-project/datahub/actions/runs/13552436222/job/37879053347?pr=12732)
when we tried an alternative action that used direct uploads.

Maybe in the future we'll revisit this, but for now it's not worth the
effort.