minor fixes/improvements in build script
akka-parsing had wrong version dependency - is part of akka-http and not akka-core, so there is no
akka-parsing of version 2.16.x. This always caused build to try to download -- only to get a 404
and then download some version that is available.

it is compressed only to be unpacked when building the docker file