fix(airflow): drop airflow 2.3 and 2.4 support
As part of https://github.com/datahub-project/datahub/pull/12982, we
need to depend on `typing-extensions` 4.5.0. The constraint files Airflow 2.3 and 2.4 require 4.4.0 or older.

I was considering reverting that change, but decided that we should be
dropping support for Airflow <2.5 anyways. We haven't been compatible
with the constraints for those for a while, and our telemetry shows that
they don't have much usage anymore (just a handful of stragglers
representing ~0.5% of our Airflow plugin users).