fix(lineage): lineage incorrect for some entities (#5427)
The getEntitiesWithLineageToEntityType function is used to get a list of
related entities to use as a filter when computing the lineage. This list was
considering only only those entities that are directly related to the entity
whose lineage is being computed. This needs to expand to entities that are
indirectly related via more than one hop.