refactor(query_combiner): automatically categorize queries as safe to combine
Removes a lot of manual function identification, stack trace iterating, etc.,
and allows even more queries to be combined. For instance, `median` is a simple
aggregate on many data warehouses, and with this new logic it will be combined
where possible, and automatically run standalone when it falls back to the ugly
`OFFSET x LIMIT 2` manual calculation inside Great Expectations.

Also, categorizing queries based on their structure rather than arbitrary names
enables the possibility of the next commit: pulling pieces out of a structured
query and combining them in a single `SELECT`, rather than a cross-join.