build(docker): use root as build context
This eliminates the copy step that copies selected files from source root to a build context that is in the project build dir. A few files like jars were copied from build dir in Dockerfiles, they need to use a path relative to repo root for the source path. This was a pattern carried over from the older docker plugin that was being used.