quickstart Reload improvements (#14982)
Adds two tasks reload and reloadEnv which work with any debug quickstart profile. They automatically detect the running profile to restart the correct set of services based on the running profile. The earlier profile specific tasks DebugReload[Env] are no longer needed (but still present, they now just delegate the call to these new tasks.

This also now handles ReloadEnv correctly for services that support hot-reloading. Some of the python service container names were not consistent across profiles, so this now detects the names based on the prefix as part of the auto-detection of running profile and services.