fix(ci): handle empty smoke-test matrix
In certain cases when no code changes are detected by ci-optimization is modified or on labels, the smoke test matrix is empty.
While it fails fast within a few seconds, the workflow is still marked as failure and not skipped. This makes the workflow status get reported as skipped instead of failed.